+++ backtrace-library/src/androidTest/java/backtraceio/library/database/BacktraceDatabaseRecordTest.java
@@ -17,12 +17,9 @@ import backtraceio.library.models.BacktraceData;
 import backtraceio.library.models.database.BacktraceDatabaseRecord;
 import backtraceio.library.models.database.BacktraceDatabaseSettings;
 import backtraceio.library.models.json.BacktraceReport;
-import backtraceio.library.services.BacktraceDatabaseContext;
 
 import static junit.framework.TestCase.assertEquals;
 import static junit.framework.TestCase.assertFalse;
-import static junit.framework.TestCase.assertNotSame;
-import static junit.framework.TestCase.assertNull;
 import static junit.framework.TestCase.assertTrue;
 
 
@@ -30,7 +27,6 @@ import static junit.framework.TestCase.assertTrue;
 public class BacktraceDatabaseRecordTest {
     private Context context;
     private String dbPath;
-    private BacktraceDatabaseContext databaseContext;
     private BacktraceDatabaseSettings databaseSettings;
     private BacktraceDatabase database;
     private String testMessage = "Example test string";
@@ -40,7 +36,6 @@ public class BacktraceDatabaseRecordTest {
         this.context = InstrumentationRegistry.getContext();
         this.dbPath = this.context.getFilesDir().getAbsolutePath();
         this.databaseSettings = new BacktraceDatabaseSettings(this.dbPath, RetryOrder.Queue);
-        this.databaseContext = new BacktraceDatabaseContext(this.context, this.databaseSettings);
         this.database = new BacktraceDatabase(this.context, dbPath);
     }
 
@@ -53,7 +48,7 @@ public class BacktraceDatabaseRecordTest {
     public void saveAndGetRecord() {
         // GIVEN
         BacktraceReport report = new BacktraceReport(testMessage);
-        BacktraceData data  = new BacktraceData(this.context, report, null);
+        BacktraceData data = new BacktraceData(this.context, report, null);
         BacktraceDatabaseRecord record = new BacktraceDatabaseRecord(data, this.dbPath);
 
         // WHEN
@@ -73,7 +68,7 @@ public class BacktraceDatabaseRecordTest {
     public void deleteFileDiagnosticPathToCorruptRecord() {
         // GIVEN
         BacktraceReport report = new BacktraceReport(testMessage);
-        BacktraceData data  = new BacktraceData(this.context, report, null);
+        BacktraceData data = new BacktraceData(this.context, report, null);
         BacktraceDatabaseRecord record = new BacktraceDatabaseRecord(data, this.dbPath);
 
         // WHEN
@@ -91,7 +86,7 @@ public class BacktraceDatabaseRecordTest {
     public void deleteFileReportPathToCorruptRecord() {
         // GIVEN
         BacktraceReport report = new BacktraceReport(testMessage);
-        BacktraceData data  = new BacktraceData(this.context, report, null);
+        BacktraceData data = new BacktraceData(this.context, report, null);
         BacktraceDatabaseRecord record = new BacktraceDatabaseRecord(data, this.dbPath);
 
         // WHEN
@@ -109,7 +104,7 @@ public class BacktraceDatabaseRecordTest {
     public void createAndDeleteRecordFiles() {
         // GIVEN
         BacktraceReport report = new BacktraceReport(testMessage);
-        BacktraceData data  = new BacktraceData(this.context, report, null);
+        BacktraceData data = new BacktraceData(this.context, report, null);
         BacktraceDatabaseRecord record = new BacktraceDatabaseRecord(data, this.dbPath);
 
         // WHEN
@@ -129,7 +124,7 @@ public class BacktraceDatabaseRecordTest {
     public void readFileAndDeserialize() {
         // GIVEN
         BacktraceReport report = new BacktraceReport(testMessage);
-        BacktraceData data  = new BacktraceData(this.context, report, null);
+        BacktraceData data = new BacktraceData(this.context, report, null);
         BacktraceDatabaseRecord record = new BacktraceDatabaseRecord(data, this.dbPath);
         record.save();