+++ backtrace-library/src/main/java/backtraceio/library/BacktraceDatabase.java
@@ -1,20 +1,22 @@
 package backtraceio.library;
 
 import android.content.Context;
-import android.util.Log;
 
 import java.io.File;
-import java.sql.Date;
+import java.util.Calendar;
 import java.util.Map;
 import java.util.Timer;
 import java.util.TimerTask;
+import java.util.concurrent.CountDownLatch;
 
 import backtraceio.library.common.FileHelper;
 import backtraceio.library.enums.database.RetryBehavior;
+import backtraceio.library.events.OnServerResponseEventListener;
 import backtraceio.library.interfaces.IBacktraceApi;
 import backtraceio.library.interfaces.IBacktraceDatabase;
 import backtraceio.library.interfaces.IBacktraceDatabaseContext;
 import backtraceio.library.interfaces.IBacktraceDatabaseFileContext;
+import backtraceio.library.logger.BacktraceLogger;
 import backtraceio.library.models.BacktraceData;
 import backtraceio.library.models.BacktraceResult;
 import backtraceio.library.models.database.BacktraceDatabaseRecord;
@@ -29,25 +31,15 @@ import backtraceio.library.services.BacktraceDatabaseFileContext;
  */
 public class BacktraceDatabase implements IBacktraceDatabase {
 
-    private IBacktraceApi BacktraceApi;
-
-    private Context _applicationContext;
-
-    private IBacktraceDatabaseContext backtraceDatabaseContext;
-
-    private IBacktraceDatabaseFileContext backtraceDatabaseFileContext;
-
-    private BacktraceDatabaseSettings databaseSettings;
-
-    private String getDatabasePath() {
-        return databaseSettings.getDatabasePath();
-    }
-
     private static boolean _timerBackgroundWork = false;
-
-    private boolean _enable = false;
-
     private static Timer _timer;
+    private transient final String LOG_TAG = BacktraceDatabase.class.getSimpleName();
+    private IBacktraceApi BacktraceApi;
+    private Context _applicationContext;
+    private IBacktraceDatabaseContext backtraceDatabaseContext;
+    private IBacktraceDatabaseFileContext backtraceDatabaseFileContext;
+    private BacktraceDatabaseSettings databaseSettings;
+    private boolean _enable = false;
 
     /**
      * Create disabled instance of BacktraceDatabase
@@ -74,22 +66,30 @@ public class BacktraceDatabase implements IBacktraceDatabase {
             throw new IllegalArgumentException("Database settings or application context is null");
         }
 
-        if (databaseSettings.getDatabasePath() == null || databaseSettings.getDatabasePath().isEmpty()) {
+        if (databaseSettings.getDatabasePath() == null || databaseSettings.getDatabasePath()
+                .isEmpty()) {
             throw new IllegalArgumentException("Database path is null or empty");
         }
 
-        if(!FileHelper.isFileExists(databaseSettings.getDatabasePath())){
+        if (!FileHelper.isFileExists(databaseSettings.getDatabasePath())) {
             boolean createDirs = new File(databaseSettings.getDatabasePath()).mkdirs();
-            if(!createDirs || !FileHelper.isFileExists(databaseSettings.getDatabasePath())) {
-                throw new IllegalArgumentException("Incorrect database path or application doesn't have permission to write to this path");
+            if (!createDirs || !FileHelper.isFileExists(databaseSettings.getDatabasePath())) {
+                throw new IllegalArgumentException("Incorrect database path or application " +
+                        "doesn't have permission to write to this path");
             }
         }
 
         this._applicationContext = context;
         this.databaseSettings = databaseSettings;
-        this.backtraceDatabaseContext = new BacktraceDatabaseContext(this._applicationContext, databaseSettings);
+        this.backtraceDatabaseContext = new BacktraceDatabaseContext(this._applicationContext,
+                databaseSettings);
         this.backtraceDatabaseFileContext = new BacktraceDatabaseFileContext(this.getDatabasePath(),
-                this.databaseSettings.getMaxDatabaseSize(), this.databaseSettings.getMaxRecordCount());
+                this.databaseSettings.getMaxDatabaseSize(), this.databaseSettings
+                .getMaxRecordCount());
+    }
+
+    private String getDatabasePath() {
+        return databaseSettings.getDatabasePath();
     }
 
     public void start() {
@@ -128,12 +128,20 @@ public class BacktraceDatabase implements IBacktraceDatabase {
         _timer.scheduleAtFixedRate(new TimerTask() {
             @Override
             public void run() {
-                Log.d("TIMER", new Date(System.currentTimeMillis()).toString());
-                if (backtraceDatabaseContext == null || backtraceDatabaseContext.isEmpty() || _timerBackgroundWork) {
-                    Log.d("NULL OR ANOTHER WORKS", new Date(System.currentTimeMillis()).toString());
+                String dateTimeNow = Calendar.getInstance().getTime().toString();
+                BacktraceLogger.d(LOG_TAG, "Timer - " + dateTimeNow);
+                if (backtraceDatabaseContext == null || backtraceDatabaseContext.isEmpty()) {
+                    BacktraceLogger.d(LOG_TAG, "Timer - Database context is null or empty: " +
+                            dateTimeNow);
                     return;
                 }
-                Log.d("TIMER CONTINUE WORKING", new Date(System.currentTimeMillis()).toString());
+
+                if (_timerBackgroundWork) {
+                    BacktraceLogger.d(LOG_TAG, "Timer - another timer works now: " + dateTimeNow);
+                    return;
+                }
+
+                BacktraceLogger.d(LOG_TAG, "Timer - continue working: " + dateTimeNow);
                 _timerBackgroundWork = true;
                 _timer.cancel();
                 _timer.purge();
@@ -141,21 +149,43 @@ public class BacktraceDatabase implements IBacktraceDatabase {
 
                 BacktraceDatabaseRecord record = backtraceDatabaseContext.first();
                 while (record != null) {
+                    final CountDownLatch threadWaiter = new CountDownLatch(1);
                     BacktraceData backtraceData = record.getBacktraceData();
                     if (backtraceData == null || backtraceData.report == null) {
+                        BacktraceLogger.d(LOG_TAG, "Timer - backtrace data or report is null - " +
+                                "deleting record");
                         delete(record);
                     } else {
-                        BacktraceResult result = BacktraceApi.send(backtraceData);
-                        if (result.status == BacktraceResultStatus.Ok) {
-                            delete(record);
-                        } else {
-                            record.close();
-                            backtraceDatabaseContext.incrementBatchRetry();
+                        final BacktraceDatabaseRecord currentRecord = record;
+                        BacktraceApi.send(backtraceData, new OnServerResponseEventListener() {
+                            @Override
+                            public void onEvent(BacktraceResult backtraceResult) {
+                                if (backtraceResult.status == BacktraceResultStatus.Ok) {
+                                    BacktraceLogger.d(LOG_TAG, "Timer - deleting record");
+                                    delete(currentRecord);
+                                } else {
+                                    BacktraceLogger.d(LOG_TAG, "Timer - closing record");
+                                    currentRecord.close();
+                                    // backtraceDatabaseContext.incrementBatchRetry(); TODO: consider another way to remove some records after few retries
+                                }
+                                threadWaiter.countDown();
+                            }
+                        });
+                        try {
+                            threadWaiter.await();
+                        } catch (Exception ex) {
+                            BacktraceLogger.e(LOG_TAG,
+                                    "Error during waiting for result in Timer", ex
+                            );
+                        }
+                        if (currentRecord.valid() && !currentRecord.locked) {
+                            BacktraceLogger.d(LOG_TAG, "Timer - record is valid and unlocked");
                             break;
                         }
                     }
                     record = backtraceDatabaseContext.first();
                 }
+                BacktraceLogger.d(LOG_TAG, "Setup new timer");
                 _timerBackgroundWork = false;
                 setupTimer();
             }
@@ -173,7 +203,7 @@ public class BacktraceDatabase implements IBacktraceDatabase {
             BacktraceData backtraceData = record.getBacktraceData();
             this.delete(record);
             if (backtraceData != null) {
-                BacktraceApi.send(backtraceData);
+                BacktraceApi.send(backtraceData, null);
             }
             record = backtraceDatabaseContext.first();
         }
@@ -265,7 +295,7 @@ public class BacktraceDatabase implements IBacktraceDatabase {
         if (backtraceDatabaseContext.count() + 1 > databaseSettings.getMaxRecordCount() &&
                 databaseSettings.getMaxRecordCount() != 0) {
             if (!backtraceDatabaseContext.removeOldestRecord()) {
-                Log.e("Backtrace.IO", "Can't remove last record. Database size is invalid");
+                BacktraceLogger.e(LOG_TAG, "Can't remove last record. Database size is invalid");
                 return false;
             }
         }
@@ -273,7 +303,8 @@ public class BacktraceDatabase implements IBacktraceDatabase {
         if (databaseSettings.getMaxDatabaseSize() != 0 && backtraceDatabaseContext
                 .getDatabaseSize() > databaseSettings.getMaxDatabaseSize()) {
             int deletePolicyRetry = 5;
-            while (backtraceDatabaseContext.getDatabaseSize() > databaseSettings.getMaxDatabaseSize()) {
+            while (backtraceDatabaseContext.getDatabaseSize() > databaseSettings
+                    .getMaxDatabaseSize()) {
                 backtraceDatabaseContext.removeOldestRecord();
                 deletePolicyRetry--; // avoid infinity loop
                 if (deletePolicyRetry == 0) {