+++ backtrace-library/src/main/java/backtraceio/library/events/RequestHandler.java
@@ -9,6 +9,7 @@ import backtraceio.library.models.BacktraceResult;
 public interface RequestHandler {
     /**
      * Event which will be executed instead of default request to Backtrace API
+     *
      * @param data which should be send to Backtrace API
      * @return response on request
      */