+++ backtrace-library/src/main/java/backtraceio/library/interfaces/IBacktraceClient.java
@@ -1,8 +1,5 @@
 package backtraceio.library.interfaces;
 
-import android.os.AsyncTask;
-
-import backtraceio.library.models.BacktraceResult;
 import backtraceio.library.models.json.BacktraceReport;
 
 /**
@@ -11,15 +8,8 @@ import backtraceio.library.models.json.BacktraceReport;
 public interface IBacktraceClient {
     /**
      * Send new report to a Backtrace API
+     *
      * @param report data which should be send to Backtrace API
-     * @return server response
      */
-    BacktraceResult send(BacktraceReport report);
-
-    /**
-     * Send asynchronously new report to a Backtrace API
-     * @param report data which should be send to Backtrace API
-     * @return async task which send data to Backtrace API and return server response
-     */
-    AsyncTask<Void, Void, BacktraceResult> sendAsync(BacktraceReport report);
+    void send(BacktraceReport report);
 }