+++ backtrace-library/src/main/java/backtraceio/library/watchdog/BacktraceWatchdogTimeoutException.java
@@ -1,0 +1,4 @@
+package backtraceio.library.watchdog;
+
+public class BacktraceWatchdogTimeoutException extends Exception{
+}