+++ backtrace-library/src/main/java/backtraceio/library/watchdog/OnApplicationNotRespondingEvent.java
@@ -1,0 +1,5 @@
+package backtraceio.library.watchdog;
+
+public interface OnApplicationNotRespondingEvent {
+    void onEvent(BacktraceWatchdogTimeoutException exception);
+}